chore: refactor, move wifi code to components
