+++ devices/esp8266-freertos/components/houseflow/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRC_DIRS "wifi"
+                       INCLUDE_DIRS "include/houseflow"
+                       )
