+++ devices/esp8266-freertos/components/houseflow/component.mk
@@ -1,0 +1,1 @@
+COMPONENT_SRCDIRS := wifi
