+++ devices/esp8266-freertos/components/houseflow/include/houseflow/wifi.h
@@ -1,0 +1,12 @@
+//
+// Created by gbaranski on 19/12/2020.
+//
+
+#ifndef ESP8266_FREERTOS_HF_WIFI_H
+#define ESP8266_FREERTOS_HF_WIFI_H
+
+static const char *WIFI_TAG = "wifi station";
+
+void wifi_init_sta(void);
+
+#endif //ESP8266_FREERTOS_HF_WIFI_H
